Comparing Change
I Changed how the scrolling function worked. Previously the code would compare source code on the website to see if it had changed. Now it compares the number of elements on screen to determine if anything has changed. Also added a print statement to print how many items have been written to the data file while in the scroll loop